refactor: address code assessment findings
Issue #2895 — Deep object traversal: introduce intermediate variables for
chains of depth 4+ in inspect_humanoid.py, example_golf_swing.py,
golf_swing_physics_engine.py, urdf_editor_window.py, and club_data_tab.py.

Issue #2894 — Hard-coded secret patterns: update adapter docstring examples
to use os.environ lookups instead of literal placeholder strings in
anthropic_adapter.py and openai_adapter.py.

Issue #2893 — Duplicated responsibility: consolidate _find_project_root()
from models.py and model_explorer.py into shared _route_utils.py; remove
duplicate compute_gravity_on_slope from terrain_physics.py (re-export from
canonical terrain/terrain_base.py); remove duplicate clamp_torque_ndof from
physics.py (re-export from physics_base.py).

Issue #2892 — Split oversized module: extract SyngasCompressionEngine and
CompressionStage from syngas_compression_calculator.py (1161 lines) into
new _syngas_compression_engine.py (~290 lines, GUI-free). Main file reduced
to ~460 lines of pure PyQt6 presentation layer. Public API unchanged.
