+++ src/api/routes/model_explorer.py
@@ -28,6 +28,7 @@ from ..models.responses import (
     ModelExplorerResponse,
     URDFTreeNode,
 )
+from ._route_utils import find_project_root
 
 router = APIRouter()
 
@@ -41,13 +42,7 @@ _MODEL_DIRS = [
 
 def _find_project_root() -> Path:
     """Find the project root directory by looking for known markers."""
-    current = Path(__file__).resolve()
-    for parent in current.parents:
-        if (parent / "src" / "shared" / "urdf").exists():
-            return parent
-        if (parent / "pyproject.toml").exists():
-            return parent
-    return Path.cwd()
+    return find_project_root()
 
 
 def _parse_urdf_link_nodes(
