+++ src/shared/python/ai/adapters/openai_adapter.py
@@ -12,8 +12,9 @@ Cost Model:
     - Typical workflow: ~$0.50-1.00
 
 Example:
+    >>> import os
     >>> from shared.python.ai.adapters.openai_adapter import OpenAIAdapter
-    >>> adapter = OpenAIAdapter(api_key="your-api-key-here")
+    >>> adapter = OpenAIAdapter(api_key=os.environ["OPENAI_API_KEY"])
     >>> response = adapter.send_message("Analyze this swing", context, tools)
 """
 
@@ -75,7 +76,8 @@ class OpenAIAdapter(BaseAgentAdapter):
         organization: Optional organization ID.
 
     Example:
-        >>> adapter = OpenAIAdapter(api_key="your-api-key-here")
+        >>> import os
+        >>> adapter = OpenAIAdapter(api_key=os.environ["OPENAI_API_KEY"])
         >>> success, message = adapter.validate_connection()
         >>> if success:
         ...     response = adapter.send_message(
